services/svc-audio/app/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic==2.5.3
pydantic-settings==2.1.0

asyncpg==0.29.0
httpx==0.27.0
tenacity==8.2.3

python-jose[cryptography]==3.3.0
python-multipart==0.0.6

azure-core==1.29.6
azure-storage-blob==12.19.0

azure-ai-contentsafety==1.0.0